Don't link files which already exist.

+++ Rakefile
@@ -2,7 +2,9 @@ def link(file)
   from = File.expand_path(File.dirname(__FILE__))
   to   = File.expand_path('~')
 
-  sh "ln -s #{from}/#{file} #{to}/.#{file}"
+  unless File.exists("#{to}/.#{file}"))
+    sh "ln -s #{from}/#{file} #{to}/.#{file}"
+  end
 end
 
 desc 'links dotfiles to the home directory'
